Split up game page
